Aibos Config Updated
PlanetTutorial.sbc file is finalized. Sky looks correct now, and landscape bumpiness is proper. Note: had to decrease some bumpiness levels from the original as it was causing crashing when my video card was set to high in the graphics option.

Voxel materials need to seriously get changed. Update resulted in the planet having a round and flat appearance, so the texture files will need to be editied, rather significantly.